Replace built-in fonts with customizable font settings.
Including Roboto fonts by default causes too many issues with the fonts not being loaded from the resources.qrc file in the HTML preview.  Fonts linked with a qrc URL are only displayed in QtWebEngine in Windows and not in Linux.  The only other way to include built-in fonts for Linux that would show up in QtWebEngine would be a data URI, which makes the preview.html file unreadable.  Finally, Roboto is a terrible choice anyway since it doesn't support Japanese and Chinese characters.  Noto would be a better choice, but the family is much too large to include.  It's best to let the end-user choose the font they want displayed in the HTML Preview.