Draw line breaks manually; fix underlining of spaces for emphasis
Remove line break rendering hack that renders hidden characters with a transparent color except for the two line break spaces at the end of a line.  Unfortunately, this hack results in a bug when emphasized text is set to be underlined in the options.  Spaces within emphasized text were not being underlined.

To continue showing line breaks, paint a line break symbol in the paintEvent() of the MarkdownEditor class.  It looks nicer anyway than the two dots that were showing before.