[fixes #813] Fix file backup error on fresh startup
On a fresh installation of ghostwriter with no previous configuration files, the application would errantly show a file backup error message box when text is entered into the editor for the first time.  The wrong variable was being used when determining whether to show this message.  The correct variable is now used.