Remove unsupported GIF and SVG for pasting
Qt only supports reading GIF images.  It can write a GIF image through
QMovie, but we are not in the business of making moves here.  :)
SVG writing is supported through QSvgGenerator; however, what is taken
from the clipboard is in QImage format, which loses the SVG
information.  It is therefore pointless to try to save the image as
an SVG.