Fix segmentation fault when .conf file is missing
Whenever ghostwriter.conf is missing (or the last used exporter within
is empty), the app will crash with a segmentation fault due to accessing
an uninitialed pointer to the current HTML exporter in appsettings.cpp.
This commit ensures that the pointer is initialized to null before being
checked for whether obtaining an exporter was successful.
BUG: 465799